Update `editStudentProfile` and `AdminIssueInstallPayment` procedures
updated `editStudentProfile` procedure that it takes only first name,
last name, email, and address and also linked it in the backend.
Furthermore, fixed bug in `AdminIssueInstallPayment` procedure so that
it does not create more installments after being issued again by
checking if the installment that corresponds to the given payment exists
in the database or not.